# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Belgü",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/investigations/{investigation_id}": {
      "get": {
        "summary": "Get Investigation",
        "operationId": "get_investigation_api_investigations__investigation_id__get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Patch Investigation",
        "operationId": "patch_investigation_api_investigations__investigation_id__patch",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/investigations/{investigation_id}/memory": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "Memory",
        "operationId": "memory_api_investigations__investigation_id__memory_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investigations/{investigation_id}/assistant": {
      "get": {
        "tags": [
          "intelligence"
        ],
        "summary": "History",
        "operationId": "history_api_investigations__investigation_id__assistant_get",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "intelligence"
        ],
        "summary": "Ask",
        "operationId": "ask_api_investigations__investigation_id__assistant_post",
        "parameters": [
          {
            "name": "investigation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Investigation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "mime": {
            "type": "string",
            "title": "Mime"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "investigation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigation Id"
          },
          "brand_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand Id"
          },
          "submission_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sha256",
          "mime",
          "size_bytes",
          "created_at"
        ],
        "title": "ArtifactView"
      },
      "AssessmentView": {
        "properties": {
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "review"
            ],
            "title": "Priority"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "automated_verdict": {
            "type": "string",
            "enum": [
              "insufficient_evidence",
              "candidate",
              "likely_phishing"
            ],
            "title": "Automated Verdict"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "default": "assessment-v1"
          }
        },
        "type": "object",
        "required": [
          "priority",
          "reasons",
          "automated_verdict"
        ],
        "title": "AssessmentView"
      },
      "AssistantRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Message"
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Evidence Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "AssistantRequest"
      },
...
      "DecisionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "value": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed_phishing",
              "benign",
              "needs_review"
            ],
            "title": "Value"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "value",
          "note",
          "created_at"
        ],
        "title": "DecisionView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InvestigationPatch": {
        "properties": {
          "workflow": {
            "type": "string",
            "enum": [
              "open",
              "closed"
            ],
            "title": "Workflow"
          }
        },
        "type": "object",
        "required": [
          "workflow"
        ],
        "title": "InvestigationPatch"
      },
      "InvestigationView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "brand_id": {
            "type": "string",
            "title": "Brand Id"
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name"
          },
          "workflow": {
            "type": "string",
            "enum": [
              "open",
              "closed"
            ],
            "title": "Workflow"
          },
          "disposition": {
            "type": "string",
            "enum": [
              "unreviewed",
              "confirmed_phishing",
              "benign",
              "needs_review"
            ],
            "title": "Disposition"
          },
          "priority": {
            "type": "string",
            "enum": [
              "normal",
              "review"
            ],
            "title": "Priority"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "entity_count": {
            "type": "integer",
            "title": "Entity Count",
            "default": 0
          },
          "evidence_count": {
            "type": "integer",
            "title": "Evidence Count",
            "default": 0
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "demo": {
            "type": "boolean",
            "title": "Demo",
            "default": false
          },
          "submissions": {
            "items": {
              "$ref": "#/components/schemas/SubmissionView"
            },
            "type": "array",
            "title": "Submissions"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/NoteView"
            },
            "type": "array",
            "title": "Notes"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/DecisionView"
            },
            "type": "array",
            "title": "Decisions"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/ArtifactView"
            },
            "type": "array",
            "title": "Attachments"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "type": "array",
            "title": "Jobs"
          },
          "assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AssessmentView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "brand_id",
          "brand_name",
          "workflow",
          "disposition",
          "priority",
          "created_at",
          "updated_at"
        ],
        "title": "InvestigationView"
      },
...
      "JobView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "investigation_id": {
            "type": "string",
            "title": "Investigation Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "collect",
              "expand",
              "analysis",
              "capture"
            ],
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "completed",
              "partial",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "additionalProperties": true,
            "type": "object",
            "title": "Progress"
          },
          "limits": {
            "additionalProperties": true,
            "type": "object",
            "title": "Limits"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "cancel_requested": {
            "type": "boolean",
            "title": "Cancel Requested"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "investigation_id",
          "kind",
          "status",
          "progress",
          "limits",
          "truncated",
          "error",
          "cancel_requested",
          "created_at",
          "updated_at"
        ],
        "title": "JobView"
      },
...
      "NoteView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "created_at"
        ],
        "title": "NoteView"
      },
...
      "SubmissionView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "enum": [
              "customer_report",
              "analyst_discovery",
              "automated_discovery"
            ],
            "title": "Source"
          },
          "note": {
            "type": "string",
            "title": "Note"
          },
          "target": {
            "$ref": "#/components/schemas/TargetView"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "note",
          "target"
        ],
        "title": "SubmissionView"
      },
      "TargetView": {
        "properties": {
          "raw_value": {
            "type": "string",
            "title": "Raw Value"
          },
          "kind": {
            "type": "string",
            "enum": [
              "url",
              "domain",
              "ip",
              "cert",
              "js_hash",
              "favicon_hash"
            ],
            "title": "Kind"
          },
          "canonical_value": {
            "type": "string",
            "title": "Canonical Value"
          },
          "hostname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hostname"
          }
        },
        "type": "object",
        "required": [
          "raw_value",
          "kind",
          "canonical_value"
        ],
        "title": "TargetView"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}